Update patch set 3
Patch Set 3:

(2 comments)

I am not a YANG expert (yet), so I'm not 100% sure if we've made the list of addresses extensible in the way we want and I'm not sure if we've "dangled" it off the NodeConnector the way we want, so I've added Ed as a reviewer.

My two concerns are whether we've insisted on all observations having a MAC address by making the key the mac and to expand the timestamp to at least two timestamps and possibly a list.

Patch-set: 3
Reviewer: Gerrit User 48 <48@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0