Update patch set 7
Patch Set 7: Code-Review+1

(7 comments)

There are a few minor nits here and one significant change we should make.

The significant change is that this is brittle to transactions failing. My opinion is that, given the timing, we should merge this, file a bug, and target the bug at Helium 1 unless André wants to make a major push and fix it.

Basically, we need to add logic for what to do if a transaction fails. The simplest thing would be to retry a fixed number of times (or for a fixed amount of time) and basically give up after that and shut down the host tracker.

The key point though, is that we should not modify our local state, i.e., the hosts map, unless the write transactions are successful and if a read or write transaction fails, we still have to try to do it again or avoid failing to process events properly.

Patch-set: 7