Bug 2971: tp-id format is incorrect for host switch pair
This patch provides a clearer and more consumable format for the host tp-id.
The current convention is similar to “host:<MAC>:openflow:<port>”.  This is
particularly confusing since the "openflow:<port>" part is derived from the
attached switch, and doesn't have anything to do with the host's properties.
This proposal is to use "host:MAC" instead.  This field is propogated to the
DLUX Topology UI, and this patch was verified visually by inspecting that the
new format propagates.

Change-Id: Ia16cad2986257e1dec9ef7d596af23a658d511ec
Signed-off-by: Ryan Goulding <ryandgoulding@gmail.com>