Update patch set 1
Patch Set 1:

> Presumably we should cherry-pick this across to stable/helium,
 > right?

We caught up about this one offline a week ago, and I thought we agreed not to cherry-pick this for Helium.  For transparency, I will repeat the summation of what we discussed in this comment:
1)  This is a usability issue rather than a functional problem.
2)  This behavior is exhibited since original Helium release (.0), through SR3.  Changing it now would probably cause confusion for users.
3)  Because this is more of a usability issue, and it has been the convention since day 1, we agreed that this should come in with the next major release, Li.

If you or others believe there is a compelling reason to push this into He, I am more than happy to cherry-pick though :).

Patch-set: 1
Reviewer: Gerrit User 3635 <3635@7fc14799-209e-464c-9743-7a06c2c21a81>
Label: Verified=0