Update patch set 3
Patch Set 3:

At this fix level no. 

In general - the model is to loose. As formulated at present it allows the creation of invalid headers like:

802.1q tag .. 802.1ad tag .. Frame

There is no enforcement in the model that the q-n-q tag is first and only the second tag is single q. 

This is why this is entitled "partial fix". That should probably be fixed at java level as I do not see how to express that sanely in yang. I will probably come back and commit that separately at a later date once I get my head around on how to do that.

Patch-set: 3