BUG 5012 - adding configurable support for timing out hosts ( in interval of seconds )
This adds two new configuration parameters (host-purge-age and
host-purge-interval) in the host tracker config yang and initial config
xml file as well as the backing logic for them.

This allows for hosts and associated links to be removed from the
operational store after a specified time since the last observations.

Author:    Sai MarapaReddy <sai.marapareddy@gmail.com>
Author:    Colin Dixon <colin@colindixon.com>
Change-Id: I636420f0cedddd757b489e8e14d91fe8a46f2340
Signed-off-by: Sai MarapaReddy <sai.marapareddy@gmail.com>