Update to work with Aluminium GA
This updates pointers to Aluminum versions, adding porting changes.

We disable the odl-l2switch-ui feature, as we do not have dlux
available. It should be pretty easy to add back.

Places where we need to change logic due to Binding mapping having been
changed receive a bit more substantial changes in that:
- we use try-with-resources blocks for read transactions
- dispatch logic is updated to reduce unnecessary nesting
- a few places are also updated to use Uint types

// FIXME: audit code changes

Change-Id: Id93b64a0b14de61ca1b0a455d386ce70c49f5a5f
Signed-off-by: Matt McCormack <mmccorm1@andrew.cmu.edu>
Signed-off-by: David Arjona <darjona@yahoo.com>
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>