Rework listener integration
The idea of registering in abstract class constructor is not nice, as it
leads to SpotBugs warnings and possible initialization problems.

Split out AbstractDecoder, which is tasked with just dealing with decoding
the package, and PacketListener, which deals with
Notification(Publish)Service integration and work dispatch.

This has the benefit of removing the need to mock services in the
decoder tests.

Change-Id: I121806fe0803b92e29ab6c3524916b70bc426880
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>