Rework listener integration
The idea of registering in abstract class constructor is not nice, as it
leads to SpotBugs warnings and possible initialization problems.

Split out AbstractDecoder, which is tasked with just dealing with decoding
the package, and PacketListener, which deals with
Notification(Publish)Service integration and work dispatch.

This has a number of benefits:
- no leakage of abstract class 'this'
- centralized logic for all decoders with package-private contract with
  PacketHandlerProvider
- direct hand-off to processing Executor via more modern registration
  method
- no need to mock MD-SAL services in decoder tests
- potential for further processing pipeline improvements

Change-Id: I121806fe0803b92e29ab6c3524916b70bc426880
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>