Improve decoding contract
Our current split between canDecode() and decode() is weird in that it
relies that canDecode() validates packet shape and then decode() blindly
relies on that.

Merge these methods in a single tryDecode() method, which makes the code
flow much more natural and understandable.

While we are here, we also correct a sizing mistake in for output packet
chain.

This merging flushes out the fact that tests are passing in DTOs that do
not pass canDecode() constraints, so we fix them up.

Change-Id: Iaac0369336777aacbccf2fe408b3d7ae97d64e6b
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>